convert declaration files into ts files & fixes
