+++ src/components/EditorBase.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useCallback, useEffect, useRef } from 'react';
 import Editor, { Monaco } from '@monaco-editor/react';
-import * as monacoEditor from 'monaco-editor';
+import type { editor } from 'monaco-editor';
 import { EditorControls } from '@components/EditorControls';
 import { ConsoleOutputContainer } from '@components/ConsoleOutputContainer';
 import { STORAGE_KEYS } from '@constants/storage';
@@ -22,14 +22,14 @@ export const EditorBase = ({ language, handleCodeExecution }: EditorBaseProps) =
   const { activeSession, updateCode } = useSession();
   const isConsoleMethodsOverridden = useRef<boolean>(false);
 
-  const handleRunCode = useCallback((editor: monacoEditor.editor.IStandaloneCodeEditor) => {
+  const handleRunCode = useCallback((editor: editor.IStandaloneCodeEditor) => {
     const code = editor.getValue();
     activeSession && updateCode(code);
     setOutput([]);
     handleCodeExecution(code);
   }, [handleCodeExecution, activeSession, updateCode]);
 
-  const debouncedSetEditorContent = useDebounce((value: string, editor: monacoEditor.editor.IStandaloneCodeEditor) => {
+  const debouncedSetEditorContent = useDebounce((value: string, editor: editor.IStandaloneCodeEditor) => {
     if (autoRun) {
       handleRunCode(editor);
     }
@@ -44,7 +44,7 @@ export const EditorBase = ({ language, handleCodeExecution }: EditorBaseProps) =
     localStorage.setItem(STORAGE_KEYS.AUTO_RUN, String(autoRun));
   }, [autoRun]);
 
-  const handleEditorMount = (editor: monacoEditor.editor.IStandaloneCodeEditor, monaco: Monaco) => {
+  const handleEditorMount = (editor: editor.IStandaloneCodeEditor, monaco: Monaco) => {
     window.editor = editor;
 
     // Add command for running code
