+++ src/types/editor.ts
