+++ src/types/editorControl.ts
